Use asyncFunction and WebKitGTK 2.17.2 IDL files
Now that jsaddle has `function` and `asyncFunction`, we can use
`asyncFunction` in the `new*Async` wrappers to preserve the async
behaviour when it is the correct thing to do.  Using `function`
will run synchronously when possible (currently when using
jsaddle-wkwebview and only when the call is triggered by an
external event).

This change includes the changes resulting from updating to the
IDL files included with WebKitGTK 2.17.2.